requirements.txt
streamlit==1.24.0
numpy==1.23.5
pandas==1.5.3
tensorflow==2.13.0
keras==2.13.0
Pillow==9.4.0
matplotlib==3.6.3
